\name{austtourism}
\alias{austtourism}
\docType{data}
\title{
Australian Tourism Dataset
}
\description{
A Price and Volume bivariate dataset.
}
\usage{data("austtourism")}
\examples{
data(austtourism)
## maybe str(austtourism) ; plot(austtourism) ...
}
\details{
The data is from \url{https://robjhyndman.com/hyndsight/tourism-time-series/}. The test and train sets are combined and the dats is reformatted to xts.
}
\source{
George Athanasopoulos, Rob J. Hyndman, Haiyan Song, Doris C. Wu, 
The Tourism Forecasting Competition,
International Journal of Forecasting,
Volume 27, Issue 3,
2011,
Pages 822-844.
}
\keyword{datasets}
