\name{electricload}
\alias{electricload}
\docType{data}
\title{
Hourly Electricity Load (Greece)
}
\description{
Total electricity load in Greece in MW as published on ENTSO-E Transparency Platform.
}
\usage{data("electricload")}
\format{
  The format is:
An ‘xts’ object on 2016-10-17/2019-04-30 22:00:00 containing:
  Data: num [1:22223, 1] 3745 3702 3783 4139 4691 ...
  Indexed by objects of class: [POSIXct,POSIXt] TZ: UTC
  xts Attributes:
 NULL
}
\details{
Hourly series of electricity load in Greece between 2016-10-17/2019-04-30.
}
\source{
Open Power System Data \url{https://doi.org/10.25832/time_series/2019-06-05}
}
\keyword{datasets}
